+++ inst/htmlwidgets/bindings/lfx-search-bindings.js
@@ -226,7 +226,6 @@ LeafletWidget.methods.searchOSMText = function(text) {
 
 
 LeafletWidget.methods.addSearchGoogle = function(options) {
-
   (function() {
     var map = this;
 
@@ -567,6 +566,12 @@ LeafletWidget.methods.addSearchFeatures = function(targetGroups, options) {
     map.searchControl = new L.Control.Search(options);
     map.searchControl.addTo(map);
 
+    map.searchControl.on("search:cancel", function(e) {
+      if (e.target.options.hideMarkerOnCollapse) {
+        e.target._map.removeLayer(this._markerSearch);
+      }
+    });
+
     map.searchControl.on('search:locationfound', function(e) {
       if (options.openPopup && e.layer._popup) {
         e.layer.openPopup();
@@ -590,7 +595,16 @@ LeafletWidget.methods.removeSearchFeatures = function(clearFeatures) {
 
     if (clearFeatures && map._searchFeatureGroupName) {
       map.layerManager.clearGroup(map._searchFeatureGroupName);
-      delete map._searchFeatureGroupName ;
+      delete map._searchFeatureGroupName;
+    }
+  }).call(this);
+};
+
+LeafletWidget.methods.clearSearchFeatures = function() {
+  (function() {
+    var map = this;
+    if (map.searchControl) {
+      map.removeLayer(map.searchControl._markerSearch)
     }
   }).call(this);
 };